Update api.py
Added the function to create the inventory list while setting up a new game. The next step is to checking to see if an inventory for a particular user exist before creating a new one. I think that I will just make it delete the old one while setting up a new game.