Update api.py
Added a check in the creategame function to see if the user has more than one game going on. If they do then it will fail to create the game.

I added two more classes in the model for the craft an item and added the craft an item function. It is not finished.